:herb: UPDATE: migrated main.js to Stomap.ts
