+++ app/index.html
@@ -68,7 +68,6 @@
 
     <script src="scripts/gen/stomap.js"></script>
     <!-- build:js({app,.tmp}) scripts/main.js -->
-    <script src="scripts/main.js"></script>
     <!-- endbuild -->
 
 </body>
